% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_http.R
\name{storr_http}
\alias{driver_http}
\alias{storr_http}
\alias{storr_server_available}
\title{Storr http client}
\usage{
storr_http(host = "localhost", port = 8888, mangle_key = FALSE,
  default_namespace = "objects")

driver_http(host = "localhost", port = 8888, mangle_key = FALSE)

storr_server_available(...)
}
\arguments{
\item{host}{Host of the storr server}

\item{port}{Port of the storr server}

\item{mangle_key}{As with \code{readRDS}; mangle the keys so that
URL/filename safe characters are used.  Uses \code{storr::encode64}.}

\item{default_namespace}{Default namespace (see
\code{\link{storr}}).}

\item{...}{Arguments passed to \code{driver_http}}
}
\description{
Storr http client and driver.
}

