% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{storr_server}
\alias{storr_server}
\title{Run a storr server}
\usage{
storr_server(st, host = "0.0.0.0", port = 8888, verbose = TRUE,
  loop = TRUE)
}
\arguments{
\item{st}{A storr object to serve.  The storr can be empty to
start (e.g., a new call to \code{storr::storr_environment()}}

\item{host}{A string that is a valid IPv4 address, or "0.0.0.0" to
listen on all IP addresses (which is the default)}

\item{port}{Port number to listened on.}

\item{verbose}{Print information about}

\item{loop}{Run the server in a loop (using
\code{httpuv::runServer}).  If \code{FALSE}, you will have to
handle requests yourself with \code{httpuv::service}.}
}
\description{
Create a storr server, powered by HTTP.
}
\details{
This is not fast; expect throughput on the order of 100-200
requests per second.

The API is not currently documented, though it will be once I'm
sure I've got this right.
}

